+++ clang/docs/ReleaseNotes.rst
@@ -108,9 +108,6 @@ C++ Language Changes
 - Allow single element access of GCC vector/ext_vector_type object to be
   constant expression. Supports the `V.xyzw` syntax and other tidbits
   as seen in OpenCL. Selecting multiple elements is left as a future work.
-- Implement `CWG1815 <https://wg21.link/CWG1815>`_. Support lifetime extension 
-  of temporary created by aggregate initialization using a default member
-  initializer.
 
 - Accept C++26 user-defined ``static_assert`` messages in C++11 as an extension.
 
@@ -300,6 +297,8 @@ Improvements to Clang's diagnostics
 - Clang now diagnoses dangling cases where a gsl-pointer is constructed from a gsl-owner object inside a container (#GH100384).
 
 - Clang now warns for u8 character literals used in C23 with ``-Wpre-c23-compat`` instead of ``-Wpre-c++17-compat``.
+
+- Clang now diagnoses cases where a dangling ``GSLOwner<GSLPointer>`` object is constructed, e.g. ``std::vector<string_view> v = {std::string()};`` (#GH100526).
 
 Improvements to Clang's time-trace
 ----------------------------------